requirements.txt
# === Простые зависимости (устанавливаются через pip install -r requirements.txt) ===

# Основная библиотека для транскрибации. ВАЖНО: устанавливать ПОСЛЕ CTranslate2.
faster-whisper

# Библиотеки для работы с аудио и утилиты
librosa
numpy
ffmpeg-python
tqdm
scipy


# === Сложные зависимости (ТРЕБУЮТ РУЧНОЙ УСТАНОВКИ) ===

# PyTorch - НЕОБХОДИМО установить вручную для поддержки GPU.
# Зайдите на https://pytorch.org/get-started/locally/ и скопируйте команду для вашей системы.
# Пример для CUDA 12.1:
# pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

# CTranslate2 - главная зависимость для faster-whisper.
# Устанавливается отдельной командой в зависимости от наличия GPU.
# Для GPU (NVIDIA):
# pip install ctranslate2 -f https://storage.googleapis.com/ctranslate2-wheels/torch-2.3.0/cuda-12.1.html
#
# Для CPU:
# pip install ctranslate2